+++ tests/testthat/test_prior.R
@@ -1,0 +1,11 @@
+context("Prior Function")
+
+test_that("Prior Function", {
+
+  dp <- DirichletProcessGaussian(rnorm(100))
+
+  priorF <- PriorFunction(dp)
+
+  expect_is(priorF, "function")
+  expect_is(priorF(0), "numeric")
+})